requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.4.2
future==0.17.1
grpcio==1.17.1
grpcio-tools==1.17.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
logzero==1.5.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycparser==2.19
pylint==2.2.2
pytest==4.0.2
python-telegram-bot==11.1.0
six==1.12.0
wrapt==1.10.11
